requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
pydantic[email]>=2.6,<3
pydantic-settings>=2.2,<3
python-dotenv>=1.0,<2
python-multipart>=0.0.9,<0.1
sqlalchemy>=2.0,<2.1
pymysql>=1.1,<1.2
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7,<2
requests>=2.31,<3
numpy>=1.26,<2
pandas>=2.2,<3
matplotlib>=3.9,<4
